+++ tests/integration/api/account/mod.rs
@@ -1,4 +1,5 @@
 mod email;
+mod export;
 mod overview;
 mod password;
 mod profile;